# docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: SuperCore API
  version: 1.0.0
  description: |
    # SuperCore - Universal Entity Management Platform API

    SuperCore is a revolutionary meta-platform for creating Core Banking systems through natural language.

    ## Core Concepts

    - **Oracle**: Platform consciousness that knows its identity, licenses, and capabilities
    - **Object Definitions**: Blueprints (DNA) for all entities in the system
    - **Instances**: Living entities created from object definitions
    - **Relationships**: Connections between instances forming a semantic graph
    - **Validation Rules**: Reusable business rules and compliance checks

    ## Authentication

    Currently in development. Future versions will use Logto for OAuth 2.0/OIDC.

    ## Rate Limiting

    Default: 100 requests per minute per IP

    ## Versioning

    API version is included in the URL path: `/api/v1/`

  contact:
    name: SuperCore Team
    email: dev@lbpay.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api-dev.supercore.lbpay.com/api/v1
    description: Development environment
  - url: https://api.supercore.lbpay.com/api/v1
    description: Production environment
...
paths:
...
  /assistant/chat:
    post:
      tags: [NL Assistant]
      summary: Chat with the Natural Language Assistant
      description: |
        General purpose chat endpoint for asking questions about the platform,
        getting help, or having the assistant guide you through tasks.
      operationId: assistantChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                  example: "Como faço para criar uma conta corrente?"
                context:
                  type: object
                  description: Optional context for the conversation
      responses:
        '200':
          description: Assistant response
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
                  suggestions:
                    type: array
                    items:
                      type: string
  /assistant/generate-object-definition:
    post:
      tags: [NL Assistant]
      summary: Generate object definition from natural language
      description: |
        Provide a natural language description of what you want to create,
        and the assistant will generate a complete object definition with
        schema, states, and validation rules.
      operationId: generateObjectDefinition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description]
              properties:
                description:
                  type: string
                  example: "Preciso de uma conta corrente para pessoa física com CPF, nome, saldo e limite"
      responses:
        '200':
          description: Generated object definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectDefinition'
  /assistant/refine-schema:
    post:
      tags: [NL Assistant]
      summary: Refine an existing schema
      description: |
        Provide feedback or changes you want to make to an existing schema,
        and the assistant will update it accordingly.
      operationId: refineSchema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [object_definition_id, refinement]
              properties:
                object_definition_id:
                  type: string
                  format: uuid
                refinement:
                  type: string
                  example: "Adiciona um campo de email obrigatório"
      responses:
        '200':
          description: Refined schema
          content:
            application/json:
              schema:
                type: object
                properties:
                  schema:
                    type: object

# ============================================================================
# COMPONENTS
# ============================================================================
components:
  schemas:
...
    ObjectDefinition:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: conta_corrente
        display_name:
          type: string
          example: Conta Corrente PF
        description:
          type: string
        version:
          type: integer
        schema:
          type: object
          description: JSON Schema Draft 7
        rules:
          type: array
          items:
            type: object
        states:
          type: object
        ui_hints:
          type: object
        relationships:
          type: array
        category:
          type: string
          enum: [BUSINESS_ENTITY, RULE, POLICY, INTEGRATION, LOGIC]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_active:
          type: boolean
...